mosquitto/client/pyproject.toml
[project]
name = "client"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "asyncio>=3.4.3",
    "bleak>=0.22.3",
    "paho-mqtt>=2.1.0",
]
